+++ NEWS.md
@@ -1,4 +1,20 @@
-# stringr 0.9.0.9000
+# stringr 1.0.0.9000
+
+* `fixed()`, `regex()`, and `coll()` now throw an error if you use them with
+  anything other than a plain string (#60).
+
+* `str_view()` creates an HTML widget to display regular expression 
+  matches (#96).
+
+* The replacement to `perl()` is `regex()` not `regexp()` (#61).
+
+* `str_subset()` now allows to set custom options for `fixed`
+  pattern search (#79, @gagolews).
+  
+* `str_replace()` and `str_replace_all()` now behave correctly when a
+  replacement string contains `$`s, `\\\\1`, etc. (#83, @gagolews).
+
+# stringr 1.0.0
 
 * stringr is now powered by [stringi](https://github.com/Rexamine/stringi) 
   instead of base R regular expressions. This improves unicode and support, and 
@@ -22,6 +38,10 @@
     str_replace_all(input, c("[ad]" = "!", "[cf]" = "?"))
     ```
 
+* `str_match()` now returns NA if an optional group doesn't match 
+  (previously it returned ""). This is more consistent with `str_extract()`
+  and other match failures.
+
 * New `str_subset()` keeps values that match a pattern. It's a convenient
   wrapper for `x[str_detect(x)]` (#21, @jiho).
 
@@ -38,8 +58,8 @@
   make it easier to locate the function you need.
 
 * `ignore.case(x)` has been deprecated in favour of 
-  `fixed|regexp|coll(x, ignore.case = TRUE)`, `perl(x)` has been deprecated in 
-  favour of `regexp(x)`.
+  `fixed|regex|coll(x, ignore.case = TRUE)`, `perl(x)` has been deprecated in 
+  favour of `regex(x)`.
 
 * `str_join()` is deprecated, please use `str_c()` instead.